Event plugin according to psr-4 format
