+++ assets/js/custom.js
@@ -2,10 +2,15 @@
     'use strict';
     $(function(){
         $('body').on('click','.cm-action',function(){
-            $(this).addClass("clickable");
-            var title = $('.clickable').siblings('.cm-title').text();
-            var description = $('.clickable').siblings('.cm-description').text();
-            var source=$('.clickable').parent().siblings().children().attr('src');
+            var title = $(this).siblings('.cm-title').text();
+            var description = $(this).siblings('.cm-description').text();
+            var source=$(this).parent().siblings().children().attr('src');
+            // var image ={
+            //     'href': source,
+            //     'type': 'image',
+            //     'title': title,
+            //     'description':description ,
+            // }
             const content = GLightbox({
                 elements: [
                     {
@@ -18,7 +23,6 @@
                 autoplayVideos: true,
             });
                 content.open();
-                $(this).removeClass("clickable");
         });
         
     });
