--- uninstall.php
@@ -1,4 +1,0 @@
-<?php
-if ( ! defined( 'WP_UNINSTALL_PLUGIN' ) ) {
-	exit;
-}
